+++ src/ambition/version_details.cc
@@ -30,6 +30,7 @@ std::vector<VersionUpdates> getVersionDetails(
           _("Added access to cycling through all the intermediate speeds (there are three steps from each standard speed level to the next; and the two intermediate speeds are shown as .3 and .7).  These can be accessed using the `shift+-` and `shift+=` keys and can be rebound using the `KEYEVENT_ _SPEED_ _DECREASE_ _SMALL` and `KEYEVENT_ _SPEED_ _INCREASE_ _SMALL` bindings in the config.txt file."),
           _("Added scrolling with the pg up/pg down keys to the Inn guest list and the Harbour ship list."),
           _("Added a \"too close\" message to Inns when they're too close to other Inns and so will have fewer guests."),
+          _("Show the number of units a leader is leading in the leadership status on the info pane."),
         },
       },
       {